requirements.txt
amqp==2.3.2
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
cffi==1.11.5
coverage==4.5.1
cryptography==2.3.1
Django==2.1.1
django-cors-headers==2.4.0
django-crontab==0.7.1
django-debug-toolbar==1.9.1
django-formtools==2.1
django-haystack==2.8.1
django-redis==4.9.0
django-ueditor==0.0.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
drf-haystack==1.8.4
elasticsearch==6.3.1
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
kombu==4.2.1
openpyxl==2.5.6
Pillow==5.2.0
pycparser==2.18
pycryptodomex==3.6.6
PyJWT==1.6.4
PyMySQL==0.9.2
python-alipay-sdk==1.8.0
python-dateutil==2.7.3
pytz==2018.5
redis==2.10.6
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
vine==1.1.4
